fix devices api missing fields

+++ apps/firmware-apps/fusion-server/fusion/internal/cluster/cluster.go
@@ -598,14 +598,14 @@ func (c *Cluster) isLocalNodePrimary() bool {
 	return false
 }
 
-func (c *Cluster) getKeepalivedPriority() int {
+func (c *Cluster) getKeepalivedPriority() int32 {
 	if c.vipMonitor != nil {
 		priority, err := c.vipMonitor.GetKeepalivedPriority()
 		if err != nil {
 			logging.GetLogger().Error("Failed to get keepalived priority: %v", err)
 			return 0
 		}
-		return priority
+		return int32(priority)
 	}
 	logging.GetLogger().Warn("VIP Monitor not set, cannot get keepalived priority")
 	return 0
